work mofo

+++ commands/region.js
@@ -1,16 +1,18 @@
 const id = require('../config.json');
 let admin = id.admin;
-const guild = new Discord.Guild(client, data);
 
 module.exports = {
-	name: 'ping',
-	description: 'Ping!',
+	name: 'region',
+	description: 'Change la région',
 	guildOnly: true,
 	execute(message) {
+
+		if (message.member.roles.cache.has(admin)){
 
         message.delete().catch(O_o=>{})
         guild.setRegion('london')
             .then(updated => console.log(`Updated guild region to ${updated.region}`))
-            .catch(console.error);
+			.catch(console.error);
+		} else { console.log('it broke my dude')}
 	},
 };
